refactor package ของ domain region
